refactor: improve daily quality findings
