+++ apps/web/src/app-shell/RightSidebar.tsx
@@ -14,20 +14,12 @@ import { useProjectStore } from "@/features/project/store/use-project-store";
 import {
   Check,
   Button,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger,
   Tabs,
   TabsList,
   TabsTab,
 } from "@workspace/ui";
 import {
   Play,
-  ChevronDown,
   GitPullRequest,
   GitPullRequestCreate,
   GitPullRequestClosed,
@@ -57,14 +49,15 @@ import { useGitInfoStore } from "@/features/git/store/use-git-info-store";
 import { PRPanel, type PRPanelHandle } from "@/features/github/components/PRPanel";
 import { CommitsPanel } from "@/features/github/components/CommitsPanel";
 import { ActionsPanel } from "@/features/github/components/ActionsPanel";
-import { useGitLog, type GitCommit } from "@/features/github/hooks/use-github";
 import { isWorkspaceSetupBlocking } from "@/features/workspace/lib/workspace-setup";
 import { useLayoutSettingsStore } from "@/features/settings/store/layout-settings-store";
 import { FileTreePanel } from "@/features/files/components/FileTreePanel";
 
 import { ChangeSection } from "@/app-shell/sidebar/ChangeSection";
+import { ChangesScopeMenu } from "@/app-shell/sidebar/ChangesScopeMenu";
 import { CommitActionsContainer } from "@/app-shell/sidebar/CommitActionsContainer";
 import { RightSidebarDialogs } from "@/app-shell/sidebar/RightSidebarDialogs";
+import { useRightSidebarChangesScope } from "@/app-shell/sidebar/useRightSidebarChangesScope";
 import { ReviewContextProvider } from "@/features/diff/components/review/ReviewContextProvider";
 import type { ReviewTarget } from "@/api/ws-api";
 import { ReviewActions } from "@/features/diff/components/review/ReviewActions";
@@ -99,175 +92,6 @@ const BASE_TABS: Array<{
 ];
 
 const FILES_TAB = { value: "files" as RightSidebarTab, labelKey: "common.files", Icon: FolderTree };
-
-type ChangesDiffScope = "branch" | "unstaged" | "staged" | "commit";
-
-interface ChangesScopeState {
-  key: string;
-  scope: ChangesDiffScope;
-  selectedCommitHash: string | null;
-  menuOpen: boolean;
-}
-
-function defaultChangesScopeState(key: string): ChangesScopeState {
-  return {
-    key,
-    scope: "branch",
-    selectedCommitHash: null,
-    menuOpen: false,
-  };
-}
-
-interface ChangesScopeMenuProps {
-  scope: ChangesDiffScope;
-  selectedCommitHash: string | null;
-  commits: GitCommit[];
-  loadingCommits: boolean;
-  stagedCount: number;
-  unstagedCount: number;
-  open: boolean;
-  isVisible: boolean;
-  onOpenChange: (open: boolean) => void;
-  onSelectScope: (scope: Exclude<ChangesDiffScope, "commit">) => void;
-  onSelectCommit: (commitHash: string) => void;
-}
-
-function formatCommitScopeLabel(commit: GitCommit | undefined, fallbackHash: string | null) {
-  if (commit) return commit.short_hash;
-  return fallbackHash ? fallbackHash.slice(0, 7) : null;
-}
-
-function ChangesScopeMenu({
-  scope,
-  selectedCommitHash,
-  commits,
-  loadingCommits,
-  stagedCount,
-  unstagedCount,
-  open,
-  isVisible,
-  onOpenChange,
-  onSelectScope,
-  onSelectCommit,
-}: ChangesScopeMenuProps) {
-  const t = useTranslations("AppShell.chrome");
-  const selectedCommit = commits.find((commit) => commit.hash === selectedCommitHash);
-  const label =
-    scope === "commit"
-      ? formatCommitScopeLabel(selectedCommit, selectedCommitHash) ??
-        t("rightSidebar.changes.scope.commit")
-      : scope === "branch"
-        ? t("rightSidebar.changes.scope.branch")
-        : scope === "staged"
-          ? t("rightSidebar.changes.scope.staged")
-          : t("rightSidebar.changes.scope.unstaged");
-
-  const renderTrailingCheck = (checked: boolean) =>
-    checked ? <Check className="size-3.5 shrink-0" /> : null;
-  const renderCountBadge = (count: number) =>
-    count > 0 ? (
-      <span className="rounded-full bg-muted px-2 py-0.5 text-xs font-medium tabular-nums text-muted-foreground">
-        {count}
-      </span>
-    ) : null;
-
-  return (
-    <DropdownMenu open={open} onOpenChange={onOpenChange}>
-      <DropdownMenuTrigger asChild>
-        <span
-          role="button"
-          title={t("rightSidebar.changes.selectScope")}
-          aria-label={t("rightSidebar.changes.selectScope")}
-          tabIndex={isVisible ? 0 : -1}
-          onPointerDown={(event) => {
-            event.stopPropagation();
-          }}
-          onMouseDown={(event) => {
-            event.stopPropagation();
-          }}
-          onClick={(event) => {
-            event.stopPropagation();
-          }}
-          onKeyDown={(event) => {
-            event.stopPropagation();
-          }}
-          className="flex h-full min-w-0 max-w-24 cursor-pointer items-center justify-center gap-1 border-l border-sidebar-border/60 px-2 text-xs font-medium text-muted-foreground transition-colors hover:bg-sidebar-accent hover:text-foreground"
-        >
-          <span className="truncate">{label}</span>
-          <ChevronDown className="size-3 shrink-0" />
-        </span>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-48">
-        <DropdownMenuItem
-          className="cursor-pointer"
-          onSelect={() => onSelectScope("unstaged")}
-        >
-          <span>{t("rightSidebar.changes.scope.unstaged")}</span>
-          {renderCountBadge(unstagedCount)}
-          <span className="flex-1" />
-          {renderTrailingCheck(scope === "unstaged")}
-        </DropdownMenuItem>
-        <DropdownMenuItem
-          className="cursor-pointer"
-          onSelect={() => onSelectScope("staged")}
-        >
-          <span>{t("rightSidebar.changes.scope.staged")}</span>
-          {renderCountBadge(stagedCount)}
-          <span className="flex-1" />
-          {renderTrailingCheck(scope === "staged")}
-        </DropdownMenuItem>
-        <DropdownMenuSub>
-          <DropdownMenuSubTrigger
-            className={cn(
-              "group/commit-scope cursor-pointer",
-              scope === "commit" &&
-                "[&>svg:last-child]:hidden hover:[&>svg:last-child]:block data-[state=open]:[&>svg:last-child]:block",
-            )}
-          >
-            <span className="flex-1">{t("rightSidebar.changes.scope.commit")}</span>
-            {scope === "commit" ? (
-              <Check className="size-3.5 shrink-0 group-hover/commit-scope:hidden group-data-[state=open]/commit-scope:hidden" />
-            ) : null}
-          </DropdownMenuSubTrigger>
-          <DropdownMenuSubContent className="max-h-72 w-80 overflow-y-auto">
-            {loadingCommits && commits.length === 0 ? (
-              <DropdownMenuItem disabled>
-                {t("rightSidebar.changes.loadingCommits")}
-              </DropdownMenuItem>
-            ) : commits.length === 0 ? (
-              <DropdownMenuItem disabled>
-                {t("rightSidebar.changes.noCommitsOnBranch")}
-              </DropdownMenuItem>
-            ) : (
-              commits.map((commit) => (
-                <DropdownMenuItem
-                  key={commit.hash}
-                  className="min-w-0 cursor-pointer"
-                  onSelect={() => onSelectCommit(commit.hash)}
-                >
-                  <span className="w-14 shrink-0 font-mono text-[11px] text-muted-foreground">
-                    {commit.short_hash}
-                  </span>
-                  <span className="min-w-0 flex-1 truncate">{commit.subject}</span>
-                  {renderTrailingCheck(
-                    scope === "commit" && selectedCommitHash === commit.hash,
-                  )}
-                </DropdownMenuItem>
-              ))
-            )}
-          </DropdownMenuSubContent>
-        </DropdownMenuSub>
-        <DropdownMenuItem
-          className="cursor-pointer"
-          onSelect={() => onSelectScope("branch")}
-        >
-          <span className="flex-1">{t("rightSidebar.changes.scope.branch")}</span>
-          {renderTrailingCheck(scope === "branch")}
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
 
 function buildWikiChatPrompt(
   prompt: string,
@@ -407,57 +231,42 @@ const RightSidebar: React.FC<RightSidebarProps> = () => {
     currentProjectPath &&
     (workspaceId || projectIdFromUrl)
   );
-  const changesScopeKey = `${currentProjectPath ?? ""}:${currentBranch ?? ""}`;
-  const [changesScopeState, setChangesScopeState] = useState<ChangesScopeState>(
-    () => defaultChangesScopeState(changesScopeKey),
-  );
-  const activeChangesScopeState =
-    changesScopeState.key === changesScopeKey
-      ? changesScopeState
-      : defaultChangesScopeState(changesScopeKey);
-  const changesScope = activeChangesScopeState.scope;
-  const selectedCommitHash = activeChangesScopeState.selectedCommitHash;
-  const changesScopeMenuOpen = activeChangesScopeState.menuOpen;
-  useEffect(() => {
-    resetCompareMode();
-    if (hasWorkingContext) {
-      void refreshRepositoryState({ fetchRemote: true });
-    }
-  }, [changesScopeKey, hasWorkingContext, refreshRepositoryState, resetCompareMode]);
-  const setChangesScopeMenuOpen = useCallback(
-    (open: boolean) => {
-      setChangesScopeState((current) => ({
-        ...(current.key === changesScopeKey
-          ? current
-          : defaultChangesScopeState(changesScopeKey)),
-        menuOpen: open,
-      }));
-    },
-    [changesScopeKey],
-  );
-
-  const commitLog = useGitLog({
-    repoPath: hasWorkingContext ? currentProjectPath ?? null : null,
-    branchKey: hasWorkingContext ? currentBranch ?? null : null,
+  const markCommitsVisited = useCallback(() => {
+    setHasVisitedCommits(true);
+  }, []);
+  const {
+    changesScope,
+    selectedCommitHash,
+    selectedCommitLabel,
+    emptyCompareLabel,
+    changesScopeMenuOpen,
+    setChangesScopeMenuOpen,
+    commitLog,
+    displayedComparedFiles,
+    displayedStagedFiles,
+    displayedUnstagedFiles,
+    displayedUntrackedFiles,
+    hasDisplayedChanges,
+    defaultBranchFallback,
+    handleSelectChangesScope,
+    handleSelectCommitScope,
+    handleChangesRefresh,
+  } = useRightSidebarChangesScope({
+    currentProjectPath,
+    currentBranch,
+    hasWorkingContext,
+    stagedFiles,
+    unstagedFiles,
+    untrackedFiles,
+    compareFiles,
+    compareRef,
+    gitStatus,
+    resetCompareMode,
+    refreshRepositoryState,
+    compareAgainstRef,
+    compareWorktreeChanges,
+    onVisitCommits: markCommitsVisited,
   });
-  const selectedCommit = useMemo(
-    () => commitLog.commits.find((commit) => commit.hash === selectedCommitHash),
-    [commitLog.commits, selectedCommitHash],
-  );
-
-  const displayedComparedFiles = compareFiles;
-  const displayedStagedFiles = stagedFiles;
-  const displayedUnstagedFiles = unstagedFiles;
-  const displayedUntrackedFiles = untrackedFiles;
-  const selectedCommitLabel =
-    selectedCommit?.short_hash ?? selectedCommitHash?.slice(0, 7) ?? null;
-  const emptyCompareLabel = changesScope === "commit" ? null : compareRef;
-  const hasDisplayedChanges =
-    changesScope === "branch" || changesScope === "commit"
-      ? displayedComparedFiles.length > 0
-      : changesScope === "staged"
-        ? displayedStagedFiles.length > 0
-        : displayedUnstagedFiles.length > 0 || displayedUntrackedFiles.length > 0;
   const showAgentChatSidebar = activeCenterTab === "wiki";
   const transformWikiChatPrompt = useCallback(
     (prompt: string) =>
@@ -468,62 +277,6 @@ const RightSidebar: React.FC<RightSidebarProps> = () => {
       ),
     [activeWikiPage, currentProject?.mainFilePath, currentProjectPath],
   );
-
-  const handleSelectChangesScope = useCallback(
-    (scope: Exclude<ChangesDiffScope, "commit">) => {
-      setChangesScopeState({
-        key: changesScopeKey,
-        scope,
-        selectedCommitHash: null,
-        menuOpen: false,
-      });
-
-      if (scope === "branch") {
-        resetCompareMode();
-        void refreshRepositoryState({ fetchRemote: true });
-        return;
-      }
-
-      void compareWorktreeChanges();
-    },
-    [changesScopeKey, compareWorktreeChanges, refreshRepositoryState, resetCompareMode],
-  );
-
-  const handleSelectCommitScope = useCallback(
-    (commitHash: string) => {
-      setChangesScopeState({
-        key: changesScopeKey,
-        scope: "commit",
-        selectedCommitHash: commitHash,
-        menuOpen: false,
-      });
-      setHasVisitedCommits(true);
-      void compareAgainstRef(commitHash);
-    },
-    [changesScopeKey, compareAgainstRef],
-  );
-
-  const handleChangesRefresh = useCallback(async () => {
-    if (changesScope === "commit" && selectedCommitHash) {
-      await compareAgainstRef(selectedCommitHash);
-      return;
-    }
-
-    if (changesScope === "staged" || changesScope === "unstaged") {
-      await compareWorktreeChanges();
-      return;
-    }
-
-    resetCompareMode();
-    await refreshRepositoryState({ fetchRemote: true });
-  }, [
-    changesScope,
-    compareAgainstRef,
-    compareWorktreeChanges,
-    refreshRepositoryState,
-    resetCompareMode,
-    selectedCommitHash,
-  ]);
 
   const renderNoContextMessage = (
     <div className="flex h-full flex-col items-center justify-center text-muted-foreground/50">
@@ -738,7 +491,7 @@ const RightSidebar: React.FC<RightSidebarProps> = () => {
                               })
                             : t("rightSidebar.changes.noChangesDetected")}
                         </span>
-                        {changesScope === "branch" && !compareRef && gitStatus?.default_branch ? (
+                        {defaultBranchFallback ? (
                           <Button
                             size="sm"
                             variant="outline"
@@ -748,7 +501,7 @@ const RightSidebar: React.FC<RightSidebarProps> = () => {
                             }}
                           >
                             {t("rightSidebar.changes.compareWithDefaultBranch", {
-                              branch: gitStatus.default_branch,
+                              branch: defaultBranchFallback,
                             })}
                           </Button>
                         ) : null}
